rust: almost complete poc-menu example, still has bugs.
- add NcMenuSection::new_separator() constructor.
- rename constructor method for an empty NcInput to new();
- fix notcurses_getc_blocking() function name and error logic.
- improve comments.
- change type of NCMENU_OPTION_* to u64 (like in C API)
- fix error_str![] macro.
- improve error messages.

+++ rust/src/widgets/menu/mod.rs
@@ -51,8 +51,8 @@ pub type NcMenuItem = crate::bindings::ffi::ncmenu_item;
 /// Section for [`NcMenu`]
 pub type NcMenuSection = crate::bindings::ffi::ncmenu_section;
 
-/// Bottom row (as opposed to top row)
-pub const NCMENU_OPTION_BOTTOM: u32 = crate::bindings::ffi::NCMENU_OPTION_BOTTOM;
+/// [NcMenuOptions] flag: Bottom row (as opposed to top row).
+pub const NCMENU_OPTION_BOTTOM: u64 = crate::bindings::ffi::NCMENU_OPTION_BOTTOM as u64;
 
-/// Hide the menu when not unrolled
-pub const NCMENU_OPTION_HIDING: u32 = crate::bindings::ffi::NCMENU_OPTION_HIDING;
+/// [NcMenuOptions] flag: Hide the menu when not unrolled
+pub const NCMENU_OPTION_HIDING: u64 = crate::bindings::ffi::NCMENU_OPTION_HIDING as u64;
